Refactor sqlite statements into db.rs
